restructuring filesystem as on main branch
