+++ pycode/memilio-epidata/memilio/epidata/getSimulationData.py
