+++ pycode/memilio-epidata/memilio/epidata/modifyDataframeSeries.py
