+++ pycode/memilio-epidata/memilio/epidata_test/test_epidata_modifyDataframeSeries.py
